Renamed Activity.model.ts to keep conventions clear.
