+++ client-app/src/models/Activity.model.ts
